Refactor Calendar components
