+++ src/App.js
@@ -2,10 +2,10 @@ import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
 import { AuthProvider } from "./context/AuthContext.js";
+import Calendar from "./pages/Calendar.jsx";
 import Home from "./pages/Home.jsx";
 import SignIn from "./pages/SignIn.jsx";
 import SignUp from "./pages/SignUp.jsx";
-import TaskPage from "./pages/TaskPage.jsx";
 const App = () => {
   return (
     <AuthProvider>
@@ -13,7 +13,7 @@ const App = () => {
         <Routes>
           <Route path="/signup" element={<SignUp />} />
           <Route path="/signin" element={<SignIn />} />
-          <Route path="/task" element={<TaskPage />} />
+          <Route path="/calendar" element={<Calendar />} />
           <Route path="/" element={<Home />} />
         </Routes>
       </Router>
